Update config.h
Update keybindings. For executing applicaitons, use the application key (AKA
Windows Key). This increases the amount of shortcuts and makes differenciating
between the type of shortcut much easier.

Update the color scheme of dwm. Currently, dwm is a little ugly for my taste.
Updating the color scheme makes it look much nicer.

Add a shortcut to launch a terminal with htop running to easily see the current
performance of the system.